chore(refactor): extract /api/artifacts/* + /api/groups/* into routes/artifacts.ts
Second bucket of the index.ts route extraction. Same pattern as #351
(routes/sessions.ts).

server/src/routes/artifacts.ts (new) — 10 handlers:
  GET    /api/artifacts/:id/sessions
  GET    /api/artifacts
  GET    /api/artifacts/archived
  PATCH  /api/artifacts/:id
  POST   /api/artifacts/:id/restore
  POST   /api/artifacts/:id/archive
  POST   /api/artifacts/:id/icon/regenerate
  PATCH  /api/groups
  POST   /api/groups/archive
  POST   /api/plugins/:id/uninstall

server/src/index.ts: 318 lines removed, 7 added (delegation).

Out of scope (separate buckets later):
- /api/resolve-path  — URL→filesystem resolver
- /api/apps/:name/start|stop  — app runtime control
- /docs/:name, /artifacts/*  — static file serving

Net across the codebase: index.ts now 1575 lines (was 2218 pre-audit;
1886 after the sessions extraction in #351). 29% smaller from baseline.

No behavioural change.
